refactored aggregator
