+++ pkg/aggregator/k8sevents.go
@@ -1,0 +1,1 @@
+package aggregator
