+++ pkg/aggregator/k8ssources.go
@@ -1,0 +1,5 @@
+package aggregator
+
+type source[DstType any] interface {
+	pushInto(dst chan<- DstType) error
+}
